requirements.txt
alembic==1.0.0
atomicwrites==1.2.1
attrs==18.2.0
click==6.7
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
numpy==1.15.3
opencv-python==3.4.3.18
pdf2image==1.0.0
pdf2jpg==0.0.5
Pillow==5.3.0
pluggy==0.7.1
py==1.6.0
pypng==0.0.18
PyQRCode==1.2.1
pytesseract==0.2.5
pytest==3.8.2
pytest-flask==0.13.0
python-dateutil==2.7.3
python-editor==1.0.3
pyzbar==0.1.7
qrcode==6.0
qrtools==0.0.2
six==1.11.0
SQLAlchemy==1.2.11
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
